addGCBias

Computes GC content for peaks

addGCBias(object, ...)

# S4 method for RangedSummarizedExperiment
addGCBias(object,
  genome = GenomeInfoDb::genome(subject))

# S4 method for SummarizedExperiment
addGCBias(object, peaks,
  genome = GenomeInfoDb::genome(peaks))

Arguments

object
(Ranged)SummarizedExperiment

...
additional arguments

genome
BSgenome object, by defualt hg19

peaks
GenomicRanges with peaks, needed if object is SummarizedExperiment and not RangedSummarizedExperiment

Value

(Ranged)SummarizedExperiment object with new column in row metadata with the gc content of the peak in question

Methods (by class)

RangedSummarizedExperiment: method for RangedSummarizedExperiment

SummarizedExperiment: method for SummarizedExperiment

Examples

data(example_counts, package = "chromVAR")
# show example on small part of data 
subset_counts <- example_counts[1:500,]
library(BSgenome.Hsapiens.UCSC.hg19)

#> Loading required package: BSgenome

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: ‘BiocGenerics’

#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from ‘package:base’:
#> 
#>     anyDuplicated, append, as.data.frame, cbind, colMeans, colnames,
#>     colSums, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, lengths, Map, mapply, match,
#>     mget, order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rowMeans, rownames, rowSums, sapply, setdiff, sort,
#>     table, tapply, union, unique, unsplit, which, which.max, which.min

#> Loading required package: S4Vectors

#> Loading required package: stats4

#> 
#> Attaching package: ‘S4Vectors’

#> The following object is masked from ‘package:base’:
#> 
#>     expand.grid

#> Loading required package: IRanges

#> Loading required package: GenomeInfoDb

#> Loading required package: GenomicRanges

#> Loading required package: Biostrings

#> Loading required package: XVector

#> 
#> Attaching package: ‘Biostrings’

#> The following object is masked from ‘package:base’:
#> 
#>     strsplit

#> Loading required package: rtracklayer

example_counts <- addGCBias(subset_counts,
                              genome = BSgenome.Hsapiens.UCSC.hg19)