computeVariability

function to compute overall variability of motif sets across samples

computeVariability(object, bootstrap_error = TRUE, bootstrap_samples = 1000,
  bootstrap_quantiles = c(0.025, 0.975), na.rm = TRUE)

Arguments

object
output from computeDeviations

bootstrap_error
compute bootstrap confidence interval

bootstrap_samples
number of bootstrap samples to take

bootstrap_quantiles
quantiles for bootstrap

na.rm
remove NAs? default is true

Value

data.frame with columns for name, variability, bootstrap lower bound, bootstrap upper bound, raw p value, adjust p value.

Examples

# NOT RUN {
# Load very small example results from computeDeviations
data(mini_dev, package = "chromVAR")
variability <- computeVariability(mini_dev)
# }