compute_deviations

Computes deviations in chromatin accessibility across sets of annotations

compute_deviations(object, annotations, ...)

  # S4 method for SummarizedExperiment,SummarizedExperiment
compute_deviations(object,
  annotations, background_peaks = get_background_peaks(object),
  expectation = compute_expectations(object))

# S4 method for SummarizedExperiment,MatrixOrmatrix
compute_deviations(object,
  annotations, background_peaks = get_background_peaks(object),
  expectation = compute_expectations(object))

# S4 method for SummarizedExperiment,list
compute_deviations(object, annotations,
  background_peaks = get_background_peaks(object),
  expectation = compute_expectations(object))

# S4 method for SummarizedExperiment,missingOrNULL
compute_deviations(object,
  annotations, background_peaks = get_background_peaks(object),
  expectation = compute_expectations(object))

# S4 method for MatrixOrmatrix,SummarizedExperiment
compute_deviations(object,
  annotations, background_peaks, expectation = compute_expectations(object))

# S4 method for MatrixOrmatrix,MatrixOrmatrix
compute_deviations(object,
  annotations, background_peaks, expectation = compute_expectations(object))

# S4 method for MatrixOrmatrix,list
compute_deviations(object, annotations,
  background_peaks, expectation = compute_expectations(object))

# S4 method for MatrixOrmatrix,missingOrNULL
compute_deviations(object, annotations,
  background_peaks, expectation = compute_expectations(object))

Arguments

object

chromVARCounts object

annotations

chromVARAnnotations object

...

additional arguments

background_peaks

(optional) background peaks matrix computed using get_background_peaks, computed internally with default paramaters if not provided

expectation

(optional) expectations computed using compute_expectations, computed automatically if not provided

Value

chromVARDeviations-class, which inherits from SummarizedExperiment, and has two assays: deviations and deviation scores.

Details

multiprocessing using bplapply

Methods (by class)

object = SummarizedExperiment,annotations = SummarizedExperiment: object and annotations are SummarizedExperiment

object = SummarizedExperiment,annotations = MatrixOrmatrix: object is SummarizedExperiment, annotations are Matrix

object = SummarizedExperiment,annotations = list: object is SummarizedExperiment, annotations are list

object = SummarizedExperiment,annotations = missingOrNULL: object is SummarizedExperiment, annotations are missing

object = MatrixOrmatrix,annotations = SummarizedExperiment: object and annotations are SummarizedExperiment

object = MatrixOrmatrix,annotations = MatrixOrmatrix: object is SummarizedExperiment, annotations are Matrix

object = MatrixOrmatrix,annotations = list: object is SummarizedExperiment, annotations are list

object = MatrixOrmatrix,annotations = missingOrNULL: object is SummarizedExperiment, annotations are missing

See also

compute_variability, plot_variability

Examples

# Load very small example counts (already filtered)
data(mini_counts, package = "chromVAR")
motifs <- get_jaspar_motifs()[c(1,2,4,298)] # only use a few for demo 
library(motifmatchr)
library(BSgenome.Hsapiens.UCSC.hg19)
motif_ix <- match_motifs(motifs, mini_counts,
                         genome = BSgenome.Hsapiens.UCSC.hg19)

# computing deviations
dev <- compute_deviations(object = mini_counts,
                         annotations = motif_ix)