compute_variability

function to compute overall variability of motif sets across samples

compute_variability(object, bootstrap_error = TRUE,
  bootstrap_samples = 1000, bootstrap_quantiles = c(0.025, 0.975),
  na.rm = TRUE)

Arguments

object

output from compute_deviations

bootstrap_error

compute bootstrap confidence interval

bootstrap_samples

number of bootstrap samples to take

bootstrap_quantiles

quantiles for bootstrap

na.rm

remove NAs? default is true

Value

data.frame with columns for name, variability, bootstrap lower bound, bootstrap upper bound, raw p value, adjust p value.

Examples

# Load very small example counts (already filtered)
data(mini_counts, package = "chromVAR")
motifs <- get_jaspar_motifs()[c(1,2,4,298)] # only use a few for demo 
library(motifmatchr)
motif_ix <- match_motifs(motifs, mini_counts)

# computing deviations
dev <- compute_deviations(object = mini_counts,
                         annotations = motif_ix)
variability <- compute_variability(dev)