deviationsTsne

Perform tsne using bias corrected deviations to visualize either cell/sample similarity or motif/kmer/annotation similarity

deviationsTsne(object, threshold = 1.5, perplexity = if (what == "samples")
  30 else 8, max_iter = 1000, theta = 0.5, what = c("samples",
  "annotations"), shiny = FALSE)

Arguments

object
deviations result

threshold
variability threshold -- use only deviations with variability greater than threshold

perplexity
perplexity parameter for tsne

max_iter
max iterations parameter for tsne

theta
theta parameter for tsne

what
tsne for similarity of samples or annotations?

shiny
load a shiny widget that enable you to explore perplexity and variability threshold parameter?

Value

data.frame with two columns for the two dimensions of tSNE output

Examples

# Load very small example results from computeDeviations
data(mini_dev, package = "chromVAR")

tsne_res <- deviationsTsne(mini_dev, threshold = 0.8, shiny = FALSE)

#> Perplexity given too high

#> Setting perplexity to 11

# setting very low variabilitiy threshold because this is mini data set
# threshold should generally be above 1
# Use plotVariability to get a sense of an appropriate threshold