filter_peaks

function to get indices of peaks that pass filters

filter_peaks(object, min_fragments_per_peak = 1, non_overlapping = TRUE,
  ix_return = FALSE)

Arguments

object

SummarizedExperiment with matrix of fragment counts per peak per sample, as computed by get_counts

min_fragments_per_peak

minimum number of fragmints in peaks across all samples

non_overlapping

reduce peak set to non-overlapping peaks, see details

ix_return

return indices of peaks to keep instead of subsetted counts object

Value

vector of indices, representing peaks that should be kept

Details

if non_overlapping is set to true, when peaks overlap the overlapping peak with lower counts is removed

See also

get_peaks, filter_samples, get_counts

Examples

data(example_counts, package = "chromVAR")

counts_filtered <- filter_samples(example_counts, min_depth = 1500,
                                  min_in_peaks = 0.15, shiny = FALSE)
counts_filtered <- filter_peaks(example_counts)